+++ internal/middleware/limithandler/limithandler.go
@@ -1,38 +1,33 @@
 package limithandler
 
 import (
-	"github.com/grpc-ecosystem/go-grpc-middleware/tags"
 	"golang.org/x/net/context"
 	"google.golang.org/grpc"
 )
 
+// GetLockKey function defines the lock key of an RPC invocation based on its context
+type GetLockKey func(context.Context) string
+
 // LimiterMiddleware contains rate limiter state
 type LimiterMiddleware struct {
 	methodLimiters map[string]*ConcurrencyLimiter
+	getLockKey     GetLockKey
+}
+
+type wrappedStream struct {
+	grpc.ServerStream
+	info              *grpc.StreamServerInfo
+	limiterMiddleware *LimiterMiddleware
+	initial           bool
 }
 
 var maxConcurrencyPerRepoPerRPC map[string]int
 
-func getRepoPath(ctx context.Context) string {
-	tags := grpc_ctxtags.Extract(ctx)
-	ctxValue := tags.Values()["grpc.request.repoPath"]
-	if ctxValue == nil {
-		return ""
-	}
-
-	s, ok := ctxValue.(string)
-	if ok {
-		return s
-	}
-
-	return ""
-}
-
 // UnaryInterceptor returns a Unary Interceptor
 func (c *LimiterMiddleware) UnaryInterceptor() grpc.UnaryServerInterceptor {
 	return func(ctx context.Context, req interface{}, info *grpc.UnaryServerInfo, handler grpc.UnaryHandler) (interface{}, error) {
-		repoPath := getRepoPath(ctx)
-		if repoPath == "" {
+		lockKey := c.getLockKey(ctx)
+		if lockKey == "" {
 			return handler(ctx, req)
 		}
 
@@ -42,7 +37,7 @@ func (c *LimiterMiddleware) UnaryInterceptor() grpc.UnaryServerInterceptor {
 			return handler(ctx, req)
 		}
 
-		return limiter.Limit(ctx, repoPath, func() (interface{}, error) {
+		return limiter.Limit(ctx, lockKey, func() (interface{}, error) {
 			return handler(ctx, req)
 		})
 	}
@@ -51,31 +46,57 @@ func (c *LimiterMiddleware) UnaryInterceptor() grpc.UnaryServerInterceptor {
 // StreamInterceptor returns a Stream Interceptor
 func (c *LimiterMiddleware) StreamInterceptor() grpc.StreamServerInterceptor {
 	return func(srv interface{}, stream grpc.ServerStream, info *grpc.StreamServerInfo, handler grpc.StreamHandler) error {
-		ctx := stream.Context()
+		wrapper := &wrappedStream{stream, info, c, true}
+		return handler(srv, wrapper)
+	}
+}
 
-		repoPath := getRepoPath(ctx)
-		if repoPath == "" {
-			return handler(srv, stream)
-		}
-
-		limiter := c.methodLimiters[info.FullMethod]
-		if limiter == nil {
-			// No concurrency limiting
-			return handler(srv, stream)
-		}
-
-		_, err := limiter.Limit(ctx, repoPath, func() (interface{}, error) {
-			return nil, handler(srv, stream)
-		})
-
+func (w *wrappedStream) RecvMsg(m interface{}) error {
+	if err := w.ServerStream.RecvMsg(m); err != nil {
 		return err
+	}
+
+	// Only perform limiting on the first request of a stream
+	if !w.initial {
+		return nil
+	}
+
+	w.initial = false
+
+	ctx := w.Context()
+
+	lockKey := w.limiterMiddleware.getLockKey(ctx)
+	if lockKey == "" {
+		return nil
+	}
+
+	limiter := w.limiterMiddleware.methodLimiters[w.info.FullMethod]
+	if limiter == nil {
+		// No concurrency limiting
+		return nil
+	}
+
+	ready := make(chan struct{})
+	go limiter.Limit(ctx, lockKey, func() (interface{}, error) {
+		close(ready)
+		<-ctx.Done()
+		return nil, nil
+	})
+
+	select {
+	case <-ctx.Done():
+		return ctx.Err()
+	case <-ready:
+		// It's our turn!
+		return nil
 	}
 }
 
 // New creates a new rate limiter
-func New() LimiterMiddleware {
+func New(getLockKey GetLockKey) LimiterMiddleware {
 	return LimiterMiddleware{
 		methodLimiters: createLimiterConfig(),
+		getLockKey:     getLockKey,
 	}
 }